+++ core/netparams/keys.go
@@ -214,8 +214,9 @@ const (
 	// how much to scale the number of proposed blocks used for performance calculation.
 	ValidatorPerformanceScalingFactor = "validator.performance.scaling.factor"
 
-	RewardsVestingBaseRate       = "rewards.vesting.baseRate"
-	RewardVestingMinimumTransfer = "rewards.vesting.minimumTransfer"
+	RewardsVestingBaseRate        = "rewards.vesting.baseRate"
+	RewardsVestingMinimumTransfer = "rewards.vesting.minimumTransfer"
+	RewardsVestingBonusMultiplier = "rewards.vesting.benefitTiers"
 
 	// Referral program.
 	ReferralProgramMaxBenefitTiers                         = "referralProgram.maxBenefitTiers"
@@ -233,7 +234,8 @@ var Deprecated = map[string]struct{}{
 
 var AllKeys = map[string]struct{}{
 	MinEpochsInTeamForMetricRewardEligibility:                {},
-	RewardVestingMinimumTransfer:                             {},
+	RewardsVestingBonusMultiplier:                            {},
+	RewardsVestingMinimumTransfer:                            {},
 	RewardsVestingBaseRate:                                   {},
 	SpotMarketTradingEnabled:                                 {},
 	PerpsMarketTradingEnabled:                                {},