+++ CHANGELOG.md
@@ -73,6 +73,8 @@
 - [9208](https://github.com/vegaprotocol/vega/issues/9208) - Refactor referral set and teams state
 - [9204](https://github.com/vegaprotocol/vega/issues/9204) - Ensure teams names are not duplicates
 - [9080](https://github.com/vegaprotocol/vega/issues/9080) - Add support for vested and vesting account in GraphQL
+- [9147](https://github.com/vegaprotocol/vega/issues/9147) - Add reward multiplier to vesting engine
+
 
 ### 🐛 Fixes